Properly handle case with no validators.
The process of pulling PII and validating was totally skipped under the old logic, where with no validators, the for loop was skipped entirely. 

ALSO: Fix an issue where the for loop iteratives _until_ 1, meaning it only ever goes down to `min_priority==2` meaning that min_priority of 2 was previously unintentionally considered the highest priority along with priority of 1. Now, it iterates until 0, meaning the lowest value of `check_priority` is 1, so min_priority then is 1, meaning the final check runs with min priority of 1 equivalent to only the priority 1 validators.